Readme

+++ src/test/java/StatsServiceTest.java
@@ -6,8 +6,8 @@ public class StatsServiceTest {
     @Test
     public void sumAmountSales() {
         StatsService service = new StatsService();
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        long expected = 180;
+        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int expected = 180;
         int actual = service.sumAmount(sales);
         Assertions.assertEquals(expected, actual);
     }
@@ -15,8 +15,8 @@ public class StatsServiceTest {
     @Test
     public void averageSumAmount() {
         StatsService service = new StatsService();
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        long expected = 15;
+        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int expected = 15;
         int actual = service.averageSumAmount(sales);
         Assertions.assertEquals(expected, actual);
     }
@@ -24,8 +24,8 @@ public class StatsServiceTest {
     @Test
     public void monthMaximumSales() {
         StatsService service = new StatsService();
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        long expected = 8;
+        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int expected = 8;
         int actual = service.monthMaximumSales(sales);
         Assertions.assertEquals(expected, actual);
     }
@@ -33,8 +33,8 @@ public class StatsServiceTest {
     @Test
     public void minSales() {
         StatsService service = new StatsService();
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        long expected = 9;
+        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int expected = 9;
         int actual = service.minSales(sales);
         Assertions.assertEquals(expected, actual);
     }
@@ -42,8 +42,8 @@ public class StatsServiceTest {
     @Test
     public void minSalesMonth() {
         StatsService service = new StatsService();
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        long expected = 5;
+        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int expected = 5;
         int actual = service.minSalesMonth(sales);
         Assertions.assertEquals(expected, actual);
 
@@ -52,8 +52,8 @@ public class StatsServiceTest {
     @Test
     public void maxSalesMonth() {
         StatsService service = new StatsService();
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        long expected = 5;
+        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int expected = 5;
         int actual = service.maxSalesMonth(sales);
         Assertions.assertEquals(expected, actual);
